requirements.txt
asgiref==3.6.0
autopep8==2.0.1
beautifulsoup4==4.11.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
crispy-bootstrap4
django-crispy_forms 
cryptography==39.0.2
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.1.6
django-allauth==0.52.0
django-bootstrap-datepicker-plus==5.0.3
django-bootstrap-v5==1.0.11
django-bootstrap5==22.2
django-heroku==0.3.1
django-widget-tweaks==1.4.12
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.17.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
oauthlib==3.2.2
protobuf==4.22.1
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
pyparsing==3.0.9
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
whitenoise==6.3.0
